pyproject.toml
[project]
name = "docparser"
version = "0.1.0"
description = ""
authors = [
    {name = "gmarcinek",email = "grzegorz.marcinek@gmail.com"}
]
packages = [{ include = "docparser" }]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "acres (==0.2.0)",
    "certifi (==2025.1.31)",
    "charset-normalizer (==3.4.1)",
    "ci-info (==0.3.0)",
    "click (==8.1.8)",
    "colorama (==0.4.6)",
    "configobj (==5.0.9)",
    "configparser (==7.1.0)",
    "etelemetry (==0.3.1)",
    "filelock (==3.17.0)",
    "fitz (==0.0.1.dev2)",
    "httplib2 (==0.22.0)",
    "idna (==3.10)",
    "isodate (==0.6.1)",
    "looseversion (==1.3.0)",
    "lxml (==5.3.0)",
    "networkx (==3.4.2)",
    "nibabel (==5.3.2)",
    "nipype (==1.9.2)",
    "numpy (==2.2.2)",
    "packaging (==24.2)",
    "pandas (==2.2.3)",
    "pathlib (==1.0.1)",
    "pillow (==11.1.0)",
    "prov (==2.0.1)",
    "puremagic (==1.28)",
    "pydot (==3.0.4)",
    "pymupdf (==1.25.2)",
    "pyparsing (==3.2.1)",
    "python-dateutil (==2.9.0.post0)",
    "pytz (==2025.1)",
    "pyxnat (==1.6.2)",
    "rdflib (==6.3.2)",
    "requests (==2.32.3)",
    "scipy (==1.15.1)",
    "simplejson (==3.19.3)",
    "six (==1.17.0)",
    "traits (==7.0.2)",
    "typing-extensions (==4.12.2)",
    "tzdata (==2025.1)",
    "urllib3 (==2.3.0)",
    "watchdog (==6.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
python = ">=3.12,<4.0"